fix: resolve mypy errors - variable name conflict and type annotation
- Fix shoplist.py: rename best_choice to up_best_choice in 'up' strategy to avoid redefinition
- Fix weekly_plan.py: update shopping_list type in TypedDict from int to float

+++ core/weekly_plan.py
@@ -20,7 +20,7 @@ class WeeklyPlanResult(TypedDict):
 
     days: List[DayPlanEntry]
     weekly_coverage: Dict[str, float]
-    shopping_list: Dict[str, int]
+    shopping_list: Dict[str, float]
     total_cost: float
